pyproject.toml
[tool.poetry]
name = "continuous-eval"
version = "0.3.14"
description = "Open-Source Evaluation for GenAI Applications."
authors = ["Yi Zhang <yi@relari.ai>", "Pasquale Antonante <pasquale@relari.ai>"]
readme = "README.md"
license = "Apache-2.0"
packages = [{include = "continuous_eval"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-dotenv = "^1.0.0"
nltk = "^3.8.1"
rouge = "^1.0.1"
openai = "^1.3.7"
tqdm = "^4.66.1"
requests = "^2.31.0"
pyyaml = "^6.0.1"
appdirs = "^1.4.4"
munkres = "^1.1.4"
thefuzz = "^0.22.1"
tenacity = ">=8.1.0,<9.0.0"
json-repair = "^0.30.0"
sqlglot = {version = "^23.17.0"}
tiktoken = "^0.8.0"
scikit-learn = "^1.6.0"
jinja2 = "^3.1.4"
joblib = "^1.4.2"

# Optional model providers
boto3 = {version = "^1.35.77", optional = true}
azure-ai-inference = {version = "^1.0.0b6", optional = true}
anthropic = {version = "^0.40.0", optional = true}
cohere = {version = "^5.13.3", optional = true}
google-generativeai = {version = "^0.8.3", optional = true}
posthog = "^3.7.4"
dill = "^0.3.9"
cloudpickle = "^3.1.0"

[tool.poetry.extras]
bedrock = ["boto3"]
azure = ["azure-ai-inference"]
anthropic = ["anthropic"]
cohere = ["cohere"]
google = ["google-generativeai"]

[tool.poetry.group.semantic]
optional = true
[tool.poetry.group.semantic.dependencies]
torch = "^2.1.1"
transformers = "^4.35.2"
sentence-transformers = "^2.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pre-commit = "^3.6.0"
ruff = "^0.7.4"

[tool.ruff]
line-length = 80

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
